// src/app/core/models/contacts/contact-model.ts
export class Contact {
  friendName: string = '';
  firstName: string = '';
  location: string = '';
  picturePath: string = '';
  connectionID: string = '';
  showType: string = '';
  status: string = '';
  titleDesc: string = '';
  Params: string = '';
  paramsAV: string = '';
  email: string = '';
  labelText: string = '';
  nameAndID: string = '';
  fromID: string = '';
  contactID: string = '';
  showConnect: string = '';
  showFollow: string = '';
}

export class SearchResult {
  entityName: string = '';
  firstName: string = '';
  location: string = '';
  picturePath: string = '';
  entityID: string = '';
  showType: string = '';
  status: string = '';
  cityState: string = '';
  Params: string = '';
  paramsAV: string = '';
  email: string = '';
  labelText: string = '';
  nameAndID: string = '';
}