// src/app/core/models/members/profile-member.model.ts
export class MemberProfileBasicInfo {
  picturePath: string = '';
  memProfileName: string = '';
  titleDesc: string = '';
  memProfileStatus: string = '';
  memProfileGender: string = '';
  memProfileDOB: string = '';
  interestedDesc: string = '';
  memProfileLookingFor: string = '';
  currentCity: string = '';
  currentStatus: string = '';
  dobDay: string = '';
  dobMonth: string = '';
  dobYear: string = '';
  firstName: string = '';
  homeNeighborhood: string = '';
  hometown: string = '';
  interestedInType: string = '';
  joinedDate: string = '';
  lastName: string = '';
  lookingForEmployment: boolean = false;
  lookingForNetworking: boolean = false;
  lookingForPartnership: boolean = false;
  lookingForRecruitment: boolean = false;
  memberID: string = '';
  middleName: string = '';
  politicalView: string = '';
  religiousView: string = '';
  sex: string = '';
  showDOBType: string = '';
  showSexInProfile: string = '';
  getLGEntitiesCount: string = '';
  sport: string = '';
  bio: string = '';
  height: string = '';
  weight: string = '';
  leftRightHandFoot: string = '';
  preferredPosition: string = '';
  secondaryPosition: string = '';
}

export class YoutubePlayList {
  etag: string = '';
  id: string = '';
  title: string = '';
  description: string = '';
  defaultThumbnail: string = '';
  defaultThumbnailHeight: string = '';
  defaultThumbnailWidth: string = '';
}

export class YoutubeVideosList {
  etag: string = '';
  id: string = '';
  title: string = '';
  description: string = '';
  defaultThumbnail: string = '';
  defaultThumbnailHeight: string = '';
  defaultThumbnailWidth: string = '';
  publishedAt: string = '';
}

export class SportsList {
  name: string = '';
  id: string = '';
  desscription: string = '';
}